+++ Homework/src/assignment/SouthernStudentHrly.java
@@ -3,7 +3,7 @@ package assignment;
 public class SouthernStudentHrly extends Student implements HourlyEmployee{
 
 	//class fields
-	private double hoursWorked;
+	private int hoursWorked;
 	private double perHour;
 	private String id;
 	
@@ -77,7 +77,7 @@ public class SouthernStudentHrly extends Student implements HourlyEmployee{
 	
 	//sets the Student's total hours for the week
 	@Override
-	public void setTotalHoursForWeek(double hoursWorked) {
+	public void setTotalHoursForWeek(int hoursWorked) {
 		this.hoursWorked = hoursWorked;
 	}
 	
@@ -98,7 +98,7 @@ public class SouthernStudentHrly extends Student implements HourlyEmployee{
 	
 	//returns the total hours worked for the week
 	@Override
-	public double getTotalHoursForWeek() {
+	public int getTotalHoursForWeek() {
 		return hoursWorked;
 	}
